+++ Realm/ObjectStore/impl/collection_notifier.hpp
@@ -27,7 +27,6 @@
 #include <exception>
 #include <functional>
 #include <mutex>
-#include <thread>
 #include <unordered_map>
 
 namespace realm {
@@ -109,7 +108,7 @@ public:
 
     virtual void run() = 0;
     void prepare_handover();
-    bool deliver(SharedGroup&, std::exception_ptr);
+    bool deliver(Realm&, SharedGroup&, std::exception_ptr);
 
 protected:
     bool have_callbacks() const noexcept { return m_have_callbacks; }
@@ -123,9 +122,6 @@ private:
     virtual void do_prepare_handover(SharedGroup&) = 0;
     virtual bool do_deliver(SharedGroup&) { return true; }
     virtual bool do_add_required_change_info(TransactionChangeInfo&) = 0;
-
-    const std::thread::id m_thread_id = std::this_thread::get_id();
-    bool is_for_current_thread() const { return m_thread_id == std::this_thread::get_id(); }
 
     mutable std::mutex m_realm_mutex;
     std::shared_ptr<Realm> m_realm;